We have added functionality to the basic interfaces in stateSpace and probability_function.
The goal is to support building the parameter spaces up from separately defined elements,
and similarly for the corresponding (prior) probability distributions.

In the stateSpace class we have added a function "attach" which allows attaching some other
stateSpace object and thereby appending the other parameters onto the parameter list. When
parameter names are needed they must be defined before attaching.

New independent_dist_product class:
To support joining the corresponding priors we have added a new type of sampleable_probability_function
which is composed of several constituent sampleable_probability_function's each corresponding to
a subspace which has been added. There is a general form of the constructor with arguments of
the joint stateSpace and a vector of the subspace priors and a few more convenient realizations
with a list of specific numbers of sub-space priors (2 to 4).

We have also begun sketching out another new type of sampleable_probability_function to realize
the induced probability function when an existing probability function has its stateSpace
transformed by a stateSpaceTransform.  A skeleton of the new class transformed_dist is given
with comments on how to realize it, but no definitions are written for the class methods.

Other minor changes to "show()" members and other small tweaks are also incorporated.

A test realization of the code on realistic problem (with gleam) yields identical results
(modulo some changes to the diagnostic output) as the previous version.