Added pt_evolve_lpost_cut option for controling how parallel temperature chains evolve and fixed erroneous evidence ratio calculation.
-pt_evolve_lpost_cut:
	This sets a cut in how much log-posterior mis-ordering is tolerated. We
generally expect, for a mature set of chains, that the hotter chains have larger
(less negative) log-posterior values.  When that is not the case it indicates
that the hotter chain is immature in that it has not found the higher-likelihood
region that the colder chain is exploring. Setting a cut on the tolerance of this
mis-ordering will force the hotter chain even hotter and prevent mixing between
the two.  With differential evolution it can be easier for the chains to relearn
the states they have been mixing with.  This makes it more likely that chains
will move from hot to cold falling into the *same* one of multiple optima, and thus
that the set will inadequately explore the overall space.  With this we wait for
hotter chains to more independently find the high likelihood regions without in
effect following the trail blazed byt the cold chains. A good rule of thumb may be
to set this to several times less than the SNR^2, if null solutions are an issue.

-I also fixed the thermal integration evidence calculation which had never really
been completed, and we spitting out nonsense.

-Also changed a reference to the RNG so that runs with prop_1d_frac can be (maybe)
identical to earlier runs when this option is off.