Implementation of effective sample size (based on autocorrelation length) is complete for the purposes of testing at the chain.cc level. The code yields sensible results on a test problem (with simple_likelihood) which agree with comparable calculations using emcee.autocorr in python to within about 10% for long enough chains.  A version with logarithmic spacing of the lag samples is also implemented and (for dlag=1.1) give results which usually agree with the linear-scaled version to within 10% or better.  The log version performs better for long chains.
The algorithm computes autocovariances with the selected grid of lag
values for each of several windows (eg sized 20000 steps each)  with the
steps decimated by some factor (eg 20).  Then the autocorrelation length
is computed from the 1+2*sum(cov[lag]/cov[0]*dlag) over the set of lags
up to the point where two successive lag points have negative covariances.
Those negative values are not added in.

This calculation is repeated, including varying number of the several
windows (beginning at the end of the chain) and a number of windows is
selected to optimize the effective sample size, which is computed from
the total_length_of_windows / autocorr_length.

The above calculations are performed for some "feature" of the state
space which is defined by a function pointer.  Typically the features
used for this kind of caluclation would be just the parameter values
but other features might also be of interest.  Choice of features will
be an important consideration for transdimensional state spaces.

In the somewhat verbose current test implementation, the features are
just the first few parameters and a call to report on the ess is
hardcoded in the parallel_chains step routine.  A more useful user
interface (and stopping condition) will be developed in ptmcmc_sampler.