New version of the autocorrelation calc with the sample average now corresponding to the actual combined window length under consideration. The renamed function compute_autocovar_windows now also returns the feature average for each window at each lag. This is the average of the window mean and the lagged window mean.  A little algebra shows that to combine windows the covariance should be corrected by the difference between this average and the overall sample average for the set of windows being considered.
This version of the code is set up to test the result, stopping after
some point and dumping data for comparison with an external calculation.
Testing on 11 example runs of SimpleLikelihood with various lengths and
degrees of decimation, that the results between the logarithmic (fast)
and linear lag spacings agree within a few percent stdev=4%, while the
fast version agrees with a similar-length run with emcee.autocorr within
stdev=10% (for the ESS). Any biases among these sets, for the ESS measures
are a few times smaller than the stdev.