Before this point gleam tests with ptmcmc were not passing their usual test.  The issue was traced to commit 2262f38, which introduced a bug in the one version of the ptmcmc proposal selection interface. This commit fixes that issue.
After that fix the gleam test still fails after the most recent commit
bc3d1042c, which included a switch in the order of state and temperature
initialization.  Before that commit temps were not initialized until
after states.  Then in the first step the posterior density for hot
chains was thus artificially low, increasing the probability that the
first proposal is accepted.  After the first accepted proposal the
process would be identical so the effect on statistics would be
irrelevant.  The test fails though.  Nominally the recent change in
order makes the first step more correct and it is needed for applying
temp-adjusted file-based initialization when called for.  This commit
nonetheless reverses the reordering to provide a testing reference.

The next commit will adopt the new ordering again and tests will need
to be updated.