Changes to gaussian dist product class to support consistent functionality on compact parameter subspace.
Issue:
Previously, and still by default, if the stateSpace has wrap boundary
then a draw from beyond the bound will end up wrapped back into the allowable
range.  This means that the drawn distribution will not actually be consistent
with evaluated probability density.  In most applications this will not occur
because a Gaussian prior, for instance would not be appropriate for a finite
wrapped parameter domain.  In some diagnositc applications applying Gaussians
it is more likely.

Resolution:
With new optional wrap_probabiliy arguemnt set to true in constructor
then the result of evaluate() will attempt to sum up the additional probability
from wrapped images of the Gaussian.  [Caveat inv_cdf is not yet consistent in this
case!]

Future:
-Fix inv_cdf with wrap_probability=true
-Also make the results consistent in the case wrap_probability=false
 by adopting a truncated Gaussian distriubtion.